refactor: Project enhancement and refactoring
- working with generic method for csv upload
