--- src/main/java/com/andresoft/fileupload/student/StudentCsvRepresentation.java
@@ -1,23 +1,0 @@
-package com.andresoft.fileupload.student;
-
-import com.opencsv.bean.CsvBindByName;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class StudentCsvRepresentation {
-
-    @CsvBindByName(column = "firstname")
-    private String fname;
-    @CsvBindByName(column = "lastname")
-    private String lname;
-    @CsvBindByName(column = "age")
-    private int age;
-}
