+++ src/main/java/com/andresoft/fileupload/student/repository/StudentRepository.java
@@ -1,5 +1,6 @@
-package com.andresoft.fileupload.student;
+package com.andresoft.fileupload.student.repository;
 
+import com.andresoft.fileupload.student.model.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StudentRepository extends JpaRepository<Student, Integer> {
