+++ students.csv
@@ -1,4 +1,4 @@
-firstname,lastname,age
+first_name,last_name,age
 andres,pardo,20
 John,Doe,30
 Student 1,Student lastname,22
